fix(cli): always render the final assistant response (hook no longer owns it) (#256)
- Remove `if not _streaming_overlay_active(session):` gate from
  _execute_with_interrupt; render_message is now called unconditionally
  for the final assistant response in all cases.
- The streaming-UI hook was changed separately to stop painting the final
  response, making app-cli the sole render owner (no cross-component gate
  to drift).
- _streaming_overlay_active had exactly one call site (the gate just
  removed), so the function is deleted entirely.
- tests/test_overlay_active_detection.py is deleted along with the
  function it tested.
- New tests/test_always_render_final_response.py (TDD): asserts
  render_message is called regardless of streaming overlay state.
  test_render_message_called_even_when_overlay_would_be_active was RED
  on pre-fix code, GREEN after.
